refactor: Add icon sizes and update imports for consistency
