+++ src/ui/iconSizes.ts
@@ -1,0 +1,6 @@
+/** heroicons 24/outline (viewBox 24x24, stroke-width 1.5) に最適化した 4px グリッドスケール */
+export const ICON_SIZE_XS = 16;
+export const ICON_SIZE_SM = 24;
+export const ICON_SIZE_MD = 32;
+export const ICON_SIZE_LG = 48;
+export const ICON_SIZE_XL = 72;
