--- src/ui/icons.ts
@@ -1,54 +1,0 @@
-import {
-  ArrowUpIcon,
-  ChevronDownIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  ChevronUpIcon,
-  CodeBracketIcon,
-  ComputerDesktopIcon,
-  DocumentTextIcon,
-  ExclamationCircleIcon,
-  ExclamationTriangleIcon,
-  HashtagIcon,
-  InformationCircleIcon,
-  LightBulbIcon,
-  LinkIcon,
-  MagnifyingGlassIcon,
-  MagnifyingGlassPlusIcon,
-  ShareIcon,
-  XMarkIcon,
-} from '@heroicons/react/24/outline';
-import { GitHubLogo } from './icons/GitHubLogo';
-import { Hatenabookmark } from './icons/Hatenabookmark';
-import { X } from './icons/X';
-
-export {
-  ArrowUpIcon,
-  ChevronDownIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-  ChevronUpIcon,
-  CodeBracketIcon,
-  ComputerDesktopIcon,
-  DocumentTextIcon,
-  ExclamationCircleIcon,
-  ExclamationTriangleIcon,
-  GitHubLogo,
-  HashtagIcon,
-  Hatenabookmark,
-  InformationCircleIcon,
-  LightBulbIcon,
-  LinkIcon,
-  MagnifyingGlassIcon,
-  MagnifyingGlassPlusIcon,
-  ShareIcon,
-  X,
-  XMarkIcon,
-};
-
-/** heroicons 24/outline (viewBox 24x24, stroke-width 1.5) に最適化した 4px グリッドスケール */
-export const ICON_SIZE_XS = 16;
-export const ICON_SIZE_SM = 24;
-export const ICON_SIZE_MD = 32;
-export const ICON_SIZE_LG = 48;
-export const ICON_SIZE_XL = 72;
