+++ src/ui/styled/theme/tokens/spacing.ts
@@ -1,5 +1,5 @@
-import { pxToRem } from '@/ui/lib/fonts';
-import { SPACING_BASE_PX } from '@/ui/styled/constant';
+import { pxToRem } from '../../../lib/fonts';
+import { SPACING_BASE_PX } from '../../constant';
 import type { TokenValues } from './types';
 
 const FIBONACCI = [1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597, 2584] as const;
